requirements.txt
ag2==0.9.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
autogen==0.9.2
certifi==2025.6.15
charset-normalizer==3.4.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.87.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
